Split sign-in class, and add sign-in flow
